requirements.txt
flask>=3.1.0
gunicorn>=22.0.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
bcrypt>=4.2.0
python-dotenv>=1.0.1
flask-limiter>=3.8.0
flask-wtf>=1.2.1
alembic>=1.13.0
tf-keras>=2.18.0
tensorflow-cpu>=2.18.0
shap>=0.46.0
reportlab>=4.2.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.0
